Replace the exponential test case with a better formed one.
The existing one referred to bags that were not defined. I think that in
this case it is reasonable to assume an empty bag, but others do not
agree with me. The new input defines all bags and carefully avoids
overflowing an int32, since the goal here is not to check large numbers.